ref(performance): Move transaction overview into folder
As more tabs are added to the transaction summary, this folder is getting
polluted with a mix of code from the transaction overview and shared utilities.
This moves the transaction overview into its own folder similar to other tabs.
