+++ static/app/routes.tsx
@@ -1366,7 +1366,9 @@ function routes() {
             component={SafeLazyLoad}
           >
             <IndexRoute
-              componentPromise={() => import('app/views/performance/transactionSummary')}
+              componentPromise={() =>
+                import('app/views/performance/transactionSummary/transactionOverview')
+              }
               component={SafeLazyLoad}
             />
             <Route
