+++ static/app/views/performance/transactionSummary/transactionOverview/durationChart.tsx
@@ -26,7 +26,7 @@ import getDynamicText from 'app/utils/getDynamicText';
 import {Theme} from 'app/utils/theme';
 import withApi from 'app/utils/withApi';
 
-import {SpanOperationBreakdownFilter} from './filter';
+import {SpanOperationBreakdownFilter} from '../filter';
 
 const QUERY_KEYS = [
   'environment',
